Fix non-determinism issue in semantics
The HostCalls (which are Instr) were rewritten into `#gatherParams`. But `#gatherParams`
had the sort `Instrs`. This meant that a rule such as

`eei.hostCall => #gatherParams(...)`

would get an `.EmptyStmts` inserted on the LHS to resolve the sort. That
caused non-determinism in the semantics, because there would be no rule
for a host call with no `.EmptyStmts` after it, and there is another
rule in the semantics which removes redundantd `.EmptyStmts` after a
single statment.